Доработки:
- добавлено пополнение баланса через Yookassa
- рефакторинг калькулятора

+++ src/main/kotlin/su/vshk/billing/bot/dao/service/PaymentNotificationDaoService.kt
@@ -3,6 +3,7 @@ package su.vshk.billing.bot.dao.service
 import org.springframework.stereotype.Service
 import reactor.core.publisher.Mono
 import reactor.core.scheduler.Schedulers
+import reactor.kotlin.core.publisher.toMono
 import su.vshk.billing.bot.dao.model.PaymentNotificationEntity
 import su.vshk.billing.bot.dao.repository.PaymentNotificationRepository
 import java.util.*
@@ -22,9 +23,16 @@ class PaymentNotificationDaoService(
             paymentNotificationRepository.save(entity)
         }.subscribeOn(Schedulers.boundedElastic())
 
-    fun deleteById(telegramId: Long): Mono<Long> =
-        Mono.fromCallable {
-            paymentNotificationRepository.deleteById(telegramId)
-            telegramId
-        }.subscribeOn(Schedulers.boundedElastic())
+    fun removeByIdSafe(telegramId: Long): Mono<Optional<PaymentNotificationEntity>> =
+        findById(telegramId)
+            .flatMap { entityOpt ->
+                if (entityOpt.isPresent) {
+                    Mono.fromCallable {
+                        paymentNotificationRepository.deleteById(telegramId)
+                        entityOpt
+                    }.subscribeOn(Schedulers.boundedElastic())
+                } else {
+                    Optional.empty<PaymentNotificationEntity>().toMono()
+                }
+            }
 }
